talán készen vannak a repository osztályok

+++ src/main/java/org/github/turostaska/dao/IActionDao.java
@@ -30,7 +30,7 @@ public interface IActionDao {
      * @param ID The desired action's ID
      * @return The action with the given ID if it exists, otherwise an empty object
      */
-    Optional<Action> getById(int ID);
+    Optional<Action> getById(Long ID);
 
     /**
      * Overwrites an existing action with the same ID with the one given as parameter if it exists
